requirements.txt
streamlit
onnxruntime
numpy
pillow
ultralytics
gdown